move face extractor endpoint to dedicated blueprint
